feat: core-module type 및 field 추가

+++ core-module/src/main/java/com/foodgo/coremodule/report/domain/Report.java
@@ -26,29 +26,29 @@ public class Report extends BaseEntity {
     private MealType type;
 
     @Column(name = "report_total", nullable = false)
-    private Integer total;
+    private Double total;
 
     @Column(name = "report_carb", nullable = false)
-    private Integer carb;
+    private Double carb;
 
     @Column(name = "report_protein", nullable = false)
-    private Integer protein;
+    private Double protein;
 
     @Column(name = "report_fat", nullable = false)
-    private Integer fat;
+    private Double fat;
 
     @Column(name = "report_sugar", nullable = false)
-    private Integer sugar;      // 당류
+    private Double sugar;      // 당류
 
     @Column(name = "report_sodium", nullable = false)
-    private Integer sodium;     // 나트륨
+    private Double sodium;     // 나트륨
 
     @Column(name = "report_cholesterol", nullable = false)
-    private Integer cholesterol; // 콜레스테롤
+    private Double cholesterol; // 콜레스테롤
 
     @Column(name = "report_saturated_fat", nullable = false)
-    private Integer saturatedFat; // 포화지방산
+    private Double saturatedFat; // 포화지방산
 
     @Column(name = "report_trans_fat", nullable = false)
-    private Integer transFat;     // 트랜스지방
+    private Double transFat;     // 트랜스지방
 }
